# docs/swagger.yaml
basePath: /api/v1/
definitions:
  models.URL:
    properties:
      path:
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license: {}
  title: Swagger Example API
  version: "1.0"
paths:
  /shortener:
    delete:
      consumes:
      - application/json
      description: test
      parameters:
      - description: url
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.URL'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
    patch:
      consumes:
      - application/json
      description: test
      parameters:
      - description: url
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.URL'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
    post:
      consumes:
      - application/json
      description: test
      parameters:
      - description: url
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.URL'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
  /shortener/{path}:
    get:
      consumes:
      - application/json
      description: test
      parameters:
      - description: path
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
  /shorteners:
    get:
      consumes:
      - application/json
      description: test
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
swagger: "2.0"